requirements.txt
langchain
llama-index
torch
dvc
hydra-core
qdrant-client
llama-parse
unstructured
python-dotenv
jinja2
ranx
ragas
datasets
pandas
numpy
transformers
sentence-transformers
accelerate
pypdf
pyrootutils